refactor(webapp): split the pure shard-placement core out of the env wrapper
The placement test reached env.server through its import of the .server
module, which the webapp guidance forbids: env.server parses the whole
environment schema at import, so the test either fails without a complete
environment or passes on ambient values.

The file already claimed the core was pure. It now is: mintShardAssignment.ts
holds the placement decision and the injected-deps resolver and imports no
env, no clock and no database. runOpsMintShard.server.ts keeps only the
env-bound wrapper, its caches and its reporters. The test moves next to the
pure module and no longer pulls env.server into its chain.

knip now sees the wrapper as an unused file rather than an unused export,
since nothing imports it until a shard key reaches an id. Ignored by path,
with a note to drop the entry with that change.
